+++ Sources/BaseFoundationiOS/Templates/Types/AsyncThrowsHandler.swift
@@ -1,0 +1,10 @@
+//
+//  AsyncThrowsHandler.swift
+//  
+//
+//  Created by pablo borquez on 07/07/23.
+//
+
+import Foundation
+
+public typealias AsyncThrowsHandler = @Sendable () async throws -> Void