+++ Sources/BaseFoundationiOS/Protocols/CoreProtocols/FirstCaseDataSource.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by pablo borquez on 06/08/23.
+//
+
+import Foundation