+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `public init` implemented on `JSONManager`.
 - `get(for:)` implemented on `DictionaryExtension`.
 - `ValueDataSource` protocol implemented.
+- `FirstCaseDataSource` protocol implemented.
 
 # 0.10.0